Moved achievement loading to cache Added loading of completed achievements on join
